Rename app.js to server.js
